# Swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "HairTransplant APIS",
    "description" : "API DOCUMENTATION",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8080",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/goalsController/update/{id}" : {
      "put" : {
        "tags" : [ "goals-controller" ],
        "operationId" : "updateGoals",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Goals"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Goals"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/medicalHistoryController/import-excel" : {
      "post" : {
        "tags" : [ "medical-history-controller" ],
        "operationId" : "importExcel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "short" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "char" : {
                    "type" : "string"
                  },
                  "int" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "long" : {
                    "type" : "integer",
                    "format" : "int64"
                  },
                  "float" : {
                    "type" : "number",
                    "format" : "float"
                  },
                  "double" : {
                    "type" : "number",
                    "format" : "double"
                  },
                  "direct" : {
                    "type" : "boolean"
                  },
                  "readOnly" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/MedicalHistory"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/lifeStyleController/import-excel" : {
      "post" : {
        "tags" : [ "life-style-controller" ],
        "operationId" : "importExcel_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "short" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "char" : {
                    "type" : "string"
                  },
                  "int" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "long" : {
                    "type" : "integer",
                    "format" : "int64"
                  },
                  "float" : {
                    "type" : "number",
                    "format" : "float"
                  },
                  "double" : {
                    "type" : "number",
                    "format" : "double"
                  },
                  "direct" : {
                    "type" : "boolean"
                  },
                  "readOnly" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/LifeStyle"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/hairLossController/import-excel" : {
      "post" : {
        "tags" : [ "hair-loss-controller" ],
        "operationId" : "importExcel_2",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "short" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "char" : {
                    "type" : "string"
                  },
                  "int" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "long" : {
                    "type" : "integer",
                    "format" : "int64"
                  },
                  "float" : {
                    "type" : "number",
                    "format" : "float"
                  },
                  "double" : {
                    "type" : "number",
                    "format" : "double"
                  },
                  "direct" : {
                    "type" : "boolean"
                  },
                  "readOnly" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/HairLoss"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/goalsController/save" : {
      "post" : {
        "tags" : [ "goals-controller" ],
        "operationId" : "createGoals",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Goals"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Goals"
                }
              }
            }
          }
        }
      }
    },
    "/api/goalsController/import-excel" : {
      "post" : {
        "tags" : [ "goals-controller" ],
        "operationId" : "importExcel_3",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "short" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "char" : {
                    "type" : "string"
                  },
                  "int" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "long" : {
                    "type" : "integer",
                    "format" : "int64"
                  },
                  "float" : {
                    "type" : "number",
                    "format" : "float"
                  },
                  "double" : {
                    "type" : "number",
                    "format" : "double"
                  },
                  "direct" : {
                    "type" : "boolean"
                  },
                  "readOnly" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Goals"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/goalsController/export-to-excel" : {
      "post" : {
        "tags" : [ "goals-controller" ],
        "operationId" : "exportGoalsToExcel",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Goals"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string",
                    "format" : "byte"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/PersonalInformations/import-excel" : {
      "post" : {
        "tags" : [ "personal-information-controller" ],
        "operationId" : "importExcel_4",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "short" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "char" : {
                    "type" : "string"
                  },
                  "int" : {
                    "type" : "integer",
                    "format" : "int32"
                  },
                  "long" : {
                    "type" : "integer",
                    "format" : "int64"
                  },
                  "float" : {
                    "type" : "number",
                    "format" : "float"
                  },
                  "double" : {
                    "type" : "number",
                    "format" : "double"
                  },
                  "direct" : {
                    "type" : "boolean"
                  },
                  "readOnly" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/PersonalInformation"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/goalsController/{id}" : {
      "get" : {
        "tags" : [ "goals-controller" ],
        "operationId" : "getGoalsById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Goals"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/goalsController/delete/{id}" : {
      "delete" : {
        "tags" : [ "goals-controller" ],
        "operationId" : "deleteGoals",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Goals"
                }
              }
            }
          }
        }
      }
    },
    "/api/goalsController/delete/all" : {
      "delete" : {
        "tags" : [ "goals-controller" ],
        "operationId" : "deleteSelectedGoals",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "array",
                "items" : {
                  "$ref" : "#/components/schemas/Goals"
                }
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "MedicalHistory" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "preExistingConditions" : {
            "type" : "string"
          },
          "currentMedications" : {
            "type" : "string"
          },
          "allergies" : {
            "type" : "string"
          },
          "previousTransplants" : {
            "type" : "string"
          },
          "dateDataEntry" : {
            "type" : "string"
          },
          "parent" : {
            "type" : "string",
            "writeOnly" : true
          },
          "stringParent" : {
            "type" : "string"
          },
          "idParent" : {
            "type" : "string"
          }
        }
      },
      "LifeStyle" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "diet" : {
            "type" : "string"
          },
          "exercise" : {
            "type" : "string"
          },
          "alcohol" : {
            "type" : "string",
            "enum" : [ "YES", "NO" ]
          },
          "tobacco" : {
            "type" : "string",
            "enum" : [ "YES", "NO" ]
          },
          "dateDataEntry" : {
            "type" : "string"
          },
          "parent" : {
            "type" : "string",
            "writeOnly" : true
          },
          "stringParent" : {
            "type" : "string"
          },
          "idParent" : {
            "type" : "string"
          }
        }
      },
      "HairLoss" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cause" : {
            "type" : "string"
          },
          "extent" : {
            "type" : "string"
          },
          "pattern" : {
            "type" : "string"
          },
          "dateDataEntry" : {
            "type" : "string"
          },
          "parent" : {
            "type" : "string",
            "writeOnly" : true
          },
          "stringParent" : {
            "type" : "string"
          },
          "idParent" : {
            "type" : "string"
          }
        }
      },
      "Goals" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "expectations" : {
            "type" : "string"
          },
          "outcome" : {
            "type" : "string"
          },
          "budget" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dateDataEntry" : {
            "type" : "string"
          },
          "parent" : {
            "type" : "string",
            "writeOnly" : true
          },
          "stringParent" : {
            "type" : "string"
          },
          "idParent" : {
            "type" : "string"
          }
        }
      },
...
      "PersonalInformation" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "firstname" : {
            "type" : "string"
          },
          "lastname" : {
            "type" : "string"
          },
          "address" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "age" : {
            "type" : "integer",
            "format" : "int32"
          },
          "medicalHistoryList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MedicalHistory"
            }
          }
        }
      }
    }
  }
}